Revert "adbd: Avoid stuck write when offline"
This reverts commit 0eb5142fe1a7d5a4dc1fd91dc6bd478db4dd6634 [formerly 9fbb2ff864b8d2b9ae5e33286c537a46cfa4aca0].

Reason for revert:
Looks like this is breaking on 'user' build variants:

06-18 00:42:29.446  7470  7472 I adbd    : Waiting for persist.adb.tls_server.enable=1
06-18 00:42:29.446  7470  7474 I adbd    : opening control endpoint /dev/usb-ffs/adb/ep0
06-18 00:42:29.448  7470  7474 I adbd    : UsbFfsConnection constructed
06-18 00:42:29.449  7470  7476 I adbd    : UsbFfs-monitor thread spawned
06-18 00:42:29.452  7470  7476 I adbd    : USB event: FUNCTIONFS_BIND
06-18 00:42:29.514  7470  7473 I adbd    : adbd_auth: received a new framework connection
06-18 00:42:29.516  7470  7473 I adbd    : received new framework fd 18 (current = -1)
06-18 00:42:37.862  7470  7476 I adbd    : USB event: FUNCTIONFS_ENABLE
06-18 00:42:37.864  7470  7532 I adbd    : UsbFfs-worker thread spawned
06-18 00:42:43.458  7470  7470 I adbd    : UsbFfs: already offline
06-18 00:42:43.460  7470  7470 I adbd    : Calling send_auth_request...
06-18 00:42:43.461  7470  7470 I adbd    : Transport is offline

With the authentication path used on 'user' builds, t->online doesn't get set until adbd_auth_verified(), but before that send_auth_request() needs to be called to request the token but fails out of send_packet() due to t->online==0, a chicken-and-egg scenario.

The original issue this intended to fix was also fixed in a different way with Change I1118f2e2a70d13f15592eb996e7084033ed5cb9d ("adb: only submit USB writes on the worker thread.") so we should be ok to revert this one.

CRs-Fixed: 2715171
Bug: 157078255
Change-Id: I440bd85ecc2a41d19827a90047f343997a2a4ecd

Former-commit-id: c3dca3b2cb1a9a63ae451db54f257248f2f66b1e